Create MinMax() in CpuPlayer.cs; new issue templates
Created a method to implement the majority of the MinMax logic. Currently skeleton implementation.

Starting off with a .CheckWin() call to make the CpuPlayer always make a specific move if it will result in a win. For now, there is a placeholder return 1 to make the CpuPlayer always make the move, even if it won't result in a win.

New 'New Issue' templates, including 'release' and 'other'.